Add buffer for recording keys after backslash
Currently a buffer (stack) is in place which starts recording the keys after a backslash is pressed. Everything after the backslash while be saved in the stack, until either a new backslash is typed or a backspace is used, after which the last item added to the stack is removed. When the buffer is empty, no key recording is happening.